Restructure tests dir
