+++ backend/ilm/tests/test_views/__init__.py
@@ -1,0 +1,12 @@
+import unittest
+
+
+def suite():
+    """
+    Discovers all test cases within the current
+    directory and subdirectories.
+
+    Returns a unittest.TestSuite object containing
+    the discovered tests.
+    """
+    return unittest.TestLoader().discover('.', pattern="test*.py")
